requirements.txt
cookiecutter==1.7.3
django==3.2.3
djangorestframework==3.12.4
